Prevent additional packet sending if failed to send a game packet
This code will check if a game packet failed to be sent to the server, and if so, it will prevent from anymore packets to arrive to the server by marking the game as "Not Interesting".
This code DOES NOT delete the game deliberately to avoid warnings in the logs - the game will be deleted when it's completed, just like any normal game.