Removed log from critical lock scope
A log was being sent in a critical lock scope inside locking the Action Buffer - which gets locked every time actions are pushed. Sending a log takes a very long time compared to the number of actions pushed - thus need to move the log reporting scope outside of the lock scope.